requirements.txt
streamlit==1.28.1
langchain==0.1.17
langchain-openai==0.1.7
langchain-community==0.0.38
langchain-chroma==0.1.1
chromadb==0.4.18
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
urllib3
